Refactor drills unit tests structure, split huge file by drill type
