+++ frontend/src/metabase-lib/test-helpers.ts
@@ -1,11 +1,18 @@
 /* istanbul ignore file */
 
 import { createMockMetadata } from "__support__/metadata";
-import type { DatabaseId, DatasetQuery } from "metabase-types/api";
+import type {
+  DatabaseId,
+  DatasetQuery,
+  DatasetColumn,
+  RowValue,
+} from "metabase-types/api";
 import {
   createSampleDatabase,
   ORDERS_ID,
 } from "metabase-types/api/mocks/presets";
+import Question from "metabase-lib/Question";
+import type StructuredQuery from "metabase-lib/queries/StructuredQuery";
 import type Metadata from "./metadata/Metadata";
 import * as ML from "./v2";
 
@@ -122,4 +129,125 @@ export const findAggregationOperator = (
     throw new Error(`Could not find aggregation operator ${operatorShortName}`);
   }
   return operator;
+};
+
+export const getAvailableDrills = ({
+  question = Question.create({
+    metadata: SAMPLE_METADATA,
+    dataset_query: DEFAULT_QUERY,
+  }),
+  clickedColumnName,
+  stageIndex = -1,
+  columns,
+  rowValues,
+  clickType,
+}: {
+  question?: Question;
+  clickedColumnName: string;
+  stageIndex?: number;
+  columns: Record<string, DatasetColumn>;
+  rowValues: Record<string, RowValue>;
+  clickType: "cell" | "header";
+}) => {
+  const query = question._getMLv2Query();
+  const legacyQuery = question.query() as StructuredQuery;
+
+  const legacyColumns = legacyQuery.columns();
+  const column = columns[clickedColumnName];
+
+  const clickedCellValue = rowValues[clickedColumnName];
+
+  const row = legacyColumns.map(({ name }) => ({
+    col: columns[name],
+    value: rowValues[name],
+  }));
+
+  const dimensions =
+    legacyQuery.aggregations().length > 0
+      ? row
+          .filter(
+            ({ col }) =>
+              col?.source === "breakout" && col?.name !== clickedColumnName,
+          )
+          .map(({ value, col }) => ({ value, column: col }))
+      : undefined;
+
+  const drills =
+    clickType === "cell"
+      ? ML.availableDrillThrus(
+          query,
+          stageIndex,
+          column,
+          clickedCellValue,
+          row,
+          dimensions,
+        )
+      : ML.availableDrillThrus(
+          query,
+          stageIndex,
+          column,
+          undefined,
+          undefined,
+          undefined,
+        );
+
+  const drillsDisplayInfo = drills.map(drill =>
+    ML.displayInfo(query, stageIndex, drill),
+  );
+
+  return {
+    drills,
+    drillsDisplayInfo,
+
+    query,
+    column,
+    stageIndex,
+  };
+};
+
+export const getAvailableDrillByType = ({
+  question = Question.create({
+    metadata: SAMPLE_METADATA,
+    dataset_query: DEFAULT_QUERY,
+  }),
+  clickedColumnName,
+  stageIndex = -1,
+  columns,
+  rowValues,
+  clickType,
+  drillType,
+}: {
+  question?: Question;
+  clickedColumnName: string;
+  stageIndex?: number;
+  columns: Record<string, DatasetColumn>;
+  rowValues: Record<string, RowValue>;
+  clickType: "cell" | "header";
+  drillType: ML.DrillThruType;
+}) => {
+  const { drills, drillsDisplayInfo, query } = getAvailableDrills({
+    question,
+    clickedColumnName,
+    stageIndex,
+    columns,
+    rowValues,
+    clickType,
+  });
+
+  const drillIndex = drillsDisplayInfo.findIndex(
+    ({ type }) => type === drillType,
+  );
+  const drill = drills[drillIndex];
+  const drillDisplayInfo = drillsDisplayInfo[drillIndex];
+
+  if (!drill) {
+    throw new TypeError(`Failed to find ${drillType} drill`);
+  }
+
+  return {
+    drill,
+    drillDisplayInfo,
+    query,
+    stageIndex,
+  };
 };
