+++ frontend/src/metabase/visualizations/components/ClickActions/ClickActionsPopover.unit.spec.tsx
@@ -7,14 +7,6 @@ import {
   screen,
 } from "__support__/ui";
 import {
-  createOrdersCreatedAtDatasetColumn,
-  createOrdersDiscountDatasetColumn,
-  createOrdersIdDatasetColumn,
-  createOrdersProductIdDatasetColumn,
-  createOrdersQuantityDatasetColumn,
-  createOrdersTableDatasetColumns,
-  createOrdersTotalDatasetColumn,
-  createOrdersUserIdDatasetColumn,
   ORDERS,
   ORDERS_ID,
   SAMPLE_DB_ID,
@@ -24,41 +16,39 @@ import type { RegularClickAction } from "metabase/visualizations/types";
 import { getMode } from "metabase/visualizations/click-actions/lib/modes";
 import { checkNotNull } from "metabase/lib/types";
 import type {
+  DatasetColumn,
   DatasetQuery,
   Filter,
+  RowValue,
   Series,
-  StructuredDatasetQuery,
 } from "metabase-types/api";
 import registerVisualizations from "metabase/visualizations/register";
 import { POPOVER_TEST_ID } from "metabase/visualizations/click-actions/actions/ColumnFormattingAction/ColumnFormattingAction";
 import { createMockSingleSeries } from "metabase-types/api/mocks";
 import type { ClickObject } from "metabase-lib/queries/drills/types";
-import { DEFAULT_QUERY, SAMPLE_METADATA } from "metabase-lib/test-helpers";
+import { SAMPLE_METADATA } from "metabase-lib/test-helpers";
 import Question from "metabase-lib/Question";
 import type StructuredQuery from "metabase-lib/queries/StructuredQuery";
 import type Dimension from "metabase-lib/Dimension";
+import {
+  ORDERS_COLUMNS,
+  ORDERS_COLUMNS_LIST,
+  ORDERS_DATASET_QUERY,
+  ORDERS_ROW_VALUES,
+} from "metabase-lib/tests/drills-common";
 
 registerVisualizations();
-
-const ORDERS_DATASET_QUERY = DEFAULT_QUERY as StructuredDatasetQuery;
-const ORDERS_COLUMNS = createOrdersTableDatasetColumns();
-const ORDERS_ROW_VALUES = {
-  ID: "3",
-  USER_ID: "1",
-  PRODUCT_ID: "105",
-  SUBTOTAL: 52.723521442619514,
-  TAX: 2.9,
-  TOTAL: 49.206842233769756,
-  DISCOUNT: 6.416679208849759,
-  CREATED_AT: "2025-12-06T22:22:48.544+02:00",
-  QUANTITY: 2,
-};
 
 describe("ClickActionsPopover", function () {
   describe("apply click actions", () => {
     describe("ColumnFormattingAction", () => {
       it("should apply column formatting to default ORDERS question on header click", async () => {
-        const { props } = await setup();
+        const { props } = await setup({
+          clicked: {
+            column: ORDERS_COLUMNS.ID,
+            value: undefined,
+          },
+        });
 
         const gearIconButton = getIcon("gear");
         expect(gearIconButton).toBeInTheDocument();
@@ -85,7 +75,12 @@ describe("ClickActionsPopover", function () {
     describe("ColumnFilterDrill", () => {
       it("should apply ColumnFilterDrill to default ORDERS question on header click", async () => {
         const filterValue = 10;
-        const { props } = await setup();
+        const { props } = await setup({
+          clicked: {
+            column: ORDERS_COLUMNS.ID,
+            value: undefined,
+          },
+        });
 
         const filterDrill = screen.getByText("Filter by this column");
         expect(filterDrill).toBeInTheDocument();
@@ -121,7 +116,12 @@ describe("ClickActionsPopover", function () {
 
     describe("SortDrill", () => {
       it("should display proper sorting controls", async () => {
-        await setup();
+        await setup({
+          clicked: {
+            column: ORDERS_COLUMNS.ID,
+            value: undefined,
+          },
+        });
 
         const sortDesc = getIcon("arrow_down");
         expect(sortDesc).toBeInTheDocument();
@@ -148,6 +148,10 @@ describe("ClickActionsPopover", function () {
               },
             },
           }),
+          clicked: {
+            column: ORDERS_COLUMNS.ID,
+            value: undefined,
+          },
         });
 
         expect(queryIcon("arrow_up")).not.toBeInTheDocument();
@@ -157,7 +161,12 @@ describe("ClickActionsPopover", function () {
       });
 
       it("should apply SortDrill to default ORDERS question on ID column header click", async () => {
-        const { props } = await setup();
+        const { props } = await setup({
+          clicked: {
+            column: ORDERS_COLUMNS.ID,
+            value: undefined,
+          },
+        });
 
         const sortDesc = getIcon("arrow_down");
         userEvent.click(sortDesc);
@@ -192,8 +201,8 @@ describe("ClickActionsPopover", function () {
     describe("SummarizeColumnByTimeDrill", () => {
       it.each([
         {
-          column: createOrdersTotalDatasetColumn(),
-          columnName: createOrdersTotalDatasetColumn().name,
+          column: ORDERS_COLUMNS.TOTAL,
+          columnName: ORDERS_COLUMNS.TOTAL.name,
           expectedCard: {
             dataset_query: getSummarizedOverTimeResultDatasetQuery(
               ORDERS.TOTAL,
@@ -203,8 +212,8 @@ describe("ClickActionsPopover", function () {
           },
         },
         {
-          column: createOrdersQuantityDatasetColumn(),
-          columnName: createOrdersQuantityDatasetColumn().name,
+          column: ORDERS_COLUMNS.QUANTITY,
+          columnName: ORDERS_COLUMNS.QUANTITY.name,
           expectedCard: {
             dataset_query: getSummarizedOverTimeResultDatasetQuery(
               ORDERS.QUANTITY,
@@ -239,8 +248,8 @@ describe("ClickActionsPopover", function () {
     describe("FKFilterDrill", () => {
       it.each([
         {
-          column: createOrdersUserIdDatasetColumn(),
-          columnName: createOrdersUserIdDatasetColumn().name,
+          column: ORDERS_COLUMNS.USER_ID,
+          columnName: ORDERS_COLUMNS.USER_ID.name,
           cellValue: "1",
           drillTitle: "View this User's Orders",
           expectedCard: {
@@ -253,8 +262,8 @@ describe("ClickActionsPopover", function () {
           },
         },
         {
-          column: createOrdersProductIdDatasetColumn(),
-          columnName: createOrdersProductIdDatasetColumn().name,
+          column: ORDERS_COLUMNS.PRODUCT_ID,
+          columnName: ORDERS_COLUMNS.PRODUCT_ID.name,
           cellValue: "111",
           drillTitle: "View this Product's Orders",
           expectedCard: {
@@ -292,8 +301,8 @@ describe("ClickActionsPopover", function () {
     describe("QuickFilterDrill", () => {
       it.each([
         {
-          column: createOrdersTotalDatasetColumn(),
-          columnName: createOrdersTotalDatasetColumn().name,
+          column: ORDERS_COLUMNS.TOTAL,
+          columnName: ORDERS_COLUMNS.TOTAL.name,
           cellValue: ORDERS_ROW_VALUES.TOTAL,
           drillTitle: ">",
           expectedCard: {
@@ -308,8 +317,8 @@ describe("ClickActionsPopover", function () {
         },
 
         {
-          column: createOrdersCreatedAtDatasetColumn(),
-          columnName: createOrdersCreatedAtDatasetColumn().name,
+          column: ORDERS_COLUMNS.CREATED_AT,
+          columnName: ORDERS_COLUMNS.CREATED_AT.name,
           cellValue: ORDERS_ROW_VALUES.CREATED_AT,
           drillTitle: "Before",
           expectedCard: {
@@ -324,8 +333,8 @@ describe("ClickActionsPopover", function () {
         },
 
         {
-          column: createOrdersDiscountDatasetColumn(),
-          columnName: createOrdersDiscountDatasetColumn().name,
+          column: ORDERS_COLUMNS.DISCOUNT,
+          columnName: ORDERS_COLUMNS.DISCOUNT.name,
           cellValue: null,
           drillTitle: "=",
           expectedCard: {
@@ -370,17 +379,18 @@ async function setup({
     metadata: SAMPLE_METADATA,
     dataset_query: ORDERS_DATASET_QUERY,
   }),
-  clicked = {
-    column: createOrdersIdDatasetColumn(),
-    value: undefined,
-  },
+  clicked,
   settings = {},
   dimension: inputDimension,
+  columns = ORDERS_COLUMNS_LIST,
+  rowValues = ORDERS_ROW_VALUES,
 }: Partial<{
   question: Question;
-  clicked: ClickObject | undefined;
+  clicked: ClickObject;
   settings: Record<string, any>;
   dimension?: Dimension;
+  columns?: DatasetColumn[];
+  rowValues?: Record<string, RowValue>;
 }> = {}) {
   const mode = checkNotNull(getMode(question));
 
@@ -396,9 +406,7 @@ async function setup({
   };
 
   const clickActions = mode.actionsForClick(
-    {
-      ...clicked,
-    },
+    clicked,
     settings,
   ) as RegularClickAction[];
 
@@ -415,12 +423,12 @@ async function setup({
       },
       {
         data: {
-          cols: [...ORDERS_COLUMNS],
+          cols: [...columns],
           rows: [],
           requested_timezone: "UTC",
           results_timezone: "Asia/Nicosia",
           results_metadata: {
-            columns: [...ORDERS_COLUMNS],
+            columns: [...columns],
           },
         },
       },
@@ -545,7 +553,7 @@ function getQuickFilterResultDatasetQuery({
   filteredColumnId: number;
   filterOperator: "=" | "!=" | ">" | "<" | "is-null" | "not-null";
   filterColumnType: string;
-  cellValue: string | number | null | undefined;
+  cellValue: RowValue;
 }): DatasetQuery {
   const filterClause = ["is-null", "not-null"].includes(filterOperator)
     ? ([
